JAS native kernel prerequisites: 2 real decomp bugs + registerDacCallback
Fixes uncovered wiring the JASystem audio kernel to run single-threaded
under SMS_NATIVE_PLATFORM (audio arc milestone 1):

- JASDSPBuf.cpp DSPBuf::mixDSP() was MISSING its return (return process(...)).
  On PPC it 'worked' by ABI accident (callee's r3 survives); host x86-64 =
  UB, returned garbage -> SIGSEGV in vframeWork(). Same class as the
  setupThreadFunc -O2 fallthrough bug.
- DsyncFrame2 truncated bufL/bufR through u32 (dsptask.h/osdsp_task.h/
  JASDSPBuf.cpp). Correct on 32-bit GC, but on LP64 it truncates a real
  heap pointer -> segfault (verified: fault addr 0xf06cf500 = low 32 bits
  of a >32-bit ptr). Widened to uintptr_t.
- JASAiCtrl.cpp Kernel::registerDacCallback was an empty stub with no SMS
  caller; filled it (dacCallbackFunc = callback) so updateDac's existing
  call site can deliver rendered DAC buffers to the native sink.

+++ include/JSystem/osdsp_task.h
@@ -3,7 +3,9 @@
 
 #include <dolphin/types.h>
 
-void DsyncFrame2(u32, u32, u32);
+// see JSystem/dsptask.h — bufL/bufR are live host pointers, must not be
+// truncated to u32 on LP64 (matches the dsptask.h declaration actually used).
+void DsyncFrame2(u32, uintptr_t, uintptr_t);
 BOOL Dsp_Running_Check();
 void Dsp_Running_Start();
